Make tests runnable in Python 3.6
In the transition from Python 3.5 to 3.6, os.fspath was "update[d] […] to check the return value of __fspath__() to be either str or bytes."
Since in python/test/lib/main_test.py's _check_core_local, argparse.ArgumentParser is mocked and autospecced, argparse.ArgumentParser().parse_args().log_dir (and .server_id) returns a MagicMock object if not told otherwise.
In python/lib/main.py's main_default, such a log_dir (and presumably server_id as well) is passed to os.path.join, which passes it to the aforementioned os.fspath.
Therefore, test case TestMainDefault fails with a "TypeError: expected str, bytes or os.PathLike object, not MagicMock".
It follows that the log_dir (and presumably server_id as well) should be properly mocked to be a str, bytes or os.PathLike object.
My proposed solution is to mock argparse.ArgumentParser the same way it is done in test_trace.