Add go.mod, created from WORKSPACE
This go.mod file was generated from WORKSPACE using the modified
bzlcompat tool at github.com/matzf/bzlcompat:gomod, running

$ bzlcompat -mod github.com/scionproto/scion
$ go mod tidy

As indicated in the comment at the top of the go.mod file, this is a
temporary solution. In scionproto/scion:master, the procedure is to
generate the bazel dependencies from go.mod. As this would be tricky to
merge correctly (various updated dependencies, updated version of go,
...), we chose the simpler approach here -- this will be abandoned in
the next rebase of the scionlab branch anyway.