Make the DRKey SV store ticker UT more robust.
It fails ~ 1 / 2000 times otherwise. Because we don't want to add sync primitives
in the store only to be able to UT it, we split the test in two: check that the
ticker runs when the store is created, and check that when the clean function is
called it removes the expired keys.