ca: add delegating handler
Add a delegating handler that is capable of talking to the CA service
for certificate renewal.

This will be plugged in to `renewal/grpc:RenewalServer` based on a config
option in the control service in a follow-up PR.

Currently, the metrics result label values are not set properly
because there is ongoing work in refactoring the metrics for the
renewal package. This will be amended in a follow-up PR.

GitOrigin-RevId: 49ffd3a6d51a06320a927c4dccf50405917ea923