daemon/snet/showpaths: allow zero for hop latency
Allow to use zero for the latency metadata of each hop. Zero can be a sensible value for this; for example when multiple interfaces are part of the same router instance, then it makes sense to announce zero for the latency between these interfaces.

Use negative value (-1 nanosecond) to represent undefined values internally, instead.

Closes #4005

GitOrigin-RevId: 51357848a53f5ff7c196c1b426f93084ece3f2cf