Workaround for issue scionproto/4019.
This is temporary (tm).
Issue scionproto/4019 reports a problem with the BFD
packets when the border router is under heavy load.
To temporarily workaround this issue, set the default value
of the detection time multipler to 100, instead of 3.
Since most configuration in scionlab do not specify values
for BFD, the detection time will be 100 * tx_interval =
100 * 200ms = 20s .