scion-pki: allow writing CMS request to file in cert renewal
Add the capability to write the CMS request to the filesystem when running
`scion-pki certs renew` via the `--req-out` flag.

This allows us to check that we can verify the CMS signed request with
openssl which gives us higher confidence in our signature creation code.

In a testing topology, the command to verify the request would be:

```
scion-pki certs renew ... --req-out cms-req.pem

openssl cms -verify \
    --in cms-req.pem \
    --inform pem \
    --CAfile gen/certs/ISD1-ASff00_0_110.root.crt \
    --purpose any \
    > /dev/null
```

(Note, this requires openssl 1.1.1d or higher)

GitOrigin-RevId: 3564c2c7ee20ce421bea97b575c4e00a4f56c47e