scion-pki: add command to format TRC and TRC payloads
Add the `scion-pki trc format` sub-command that allows formatting a TRC
or TRC payload in either PEM or DER format.

```
'format' prints the TRC or TRC payload in a different format.

The PEM type for a TRC is 'TRC', and for a TRC payload it is 'TRC PAYLOAD'.

By default, the output is PEM encoded. DER format can be requested by providing
the 'der' in the --format flag. DER format is output as raw bytes and could contain
control sequences. Be careful when stdout is not redirecting stdout the DER format
is specified.
```

Going forward, we want to encode TRCs in PEM as this is a lot nicer
for humans to interact with. It can be displayed in the terminal without
breaking it, and copy pasting from one place to the other is easier in
PEM.

GitOrigin-RevId: 3d9d6926e60f1a0a4a6e0069566ebdc5437cf552