doc: add Authenticator End-to-End option
The Authenticator End-to-End option roughly follows what was previously
described in the SCION book as the "SCION Packet Security Extension". As
the name implies, however, this does not intend to be applicable for
packet encryption -- I think that encrypted packets can be more
appropriately represented by a separate next-header type analogous to IP
Encapsulating Security Payload.

This Authenticator option will be used for authenticated SCMP (to be described in a separate document, see #3969).

[doc]

Closes #4014

GitOrigin-RevId: 8511f3ea7e5df86eef52ca10f4f97a268c50da84