slayers: check for repetition of extension headers
The hop-by-hop and end-to-end extension headers can occur at most once
each in a scion packet, and they must be in the correct order, first
HBH, then E2E.
Both the serialization and decode logic had only checked the order
constraint, i.e. it checked that NextHdr of E2E is not HBH. There
was no check for repeated headers.

Note: this check is important not only for adherence to the protocol specifications, but also because gopacket's DecodingLayerParser cannot sensibly handle packets with repeated layer types.

Fixed in two places:
- split up the scionNextLayerType helper function into three related but
  separate parts: scionNextLayerType (for use in the SCION base packet),
  scionNextLayerTypeAfterHBH and scionNextLayerTypeAfterE2E.
  These functions now don't allow the transition back to the same layer.

  golang trick: this allows to move the registration of the layer type
  to a variable instead of init, as it breaks the initialization cycle.
  We get a compile time guarantee that none of the layers is its own
  NextLayerType.

- check NextHdr for repetition in both serialize and decode

The first part is not strictly necessary, but I think that it helps to
clarify what's going on.

Also:
- fix extension protocol identifier in some places in the docs
- fix protocol identifier string representation for HopByHop extension

Closes #4049

GitOrigin-RevId: 30b11c248cdfbd61a46ac13ac29da72d06818516