squic: offload accepting QUIC streams into the net.Conn wrapper
Until now, we have accepted the QUIC stream inside our QUIC listener
and then wrapped the stream as a net.Conn.

Attached to it was this seemingly innocent comment:

```go
// Accept accepts the first stream on a session and wraps it as a net.Conn.
//
// XXX(roosd): Accept blocks until the first bytes on the stream are received.
// This will limit QPS heavily, but we should not yet be in a range where this
// matters too much.
```

It turns out, QPS is truly heavily limited by this behavior.

With this change, accepting the stream is offloaded to the net.Conn wrapper.
On the first invocation of `Read` or `Write`, the implementation attempts
to accept the stream. Accepting the stream is still limited by the
default 5 second timeout as before.

With this change, the accept loop can continue without waiting for the
streams to be accepted. This is especially important, because the QUIC
library only supports up to 32 open QUIC sessions that have not been accepted yet
at the same time.

GitOrigin-RevId: 5699ffeb15932b39661f9604157f8111db4fb474