tools: add topodot
A simple script `topodot` to visualize the AS topology represented by `*.topo`-files using `dot`.
By default, this will output a description of the graph in the DOT language. When the `--show` option is passed, it will directly call dot to render the graph and open the resulting image file.

Implementation note: this doesn't use any graphviz python library as these don't really add much abstraction and it would just be an unnecessary additional dependency to deal with.

I've had this lying around for a rather long time and used it occasionally to find my way through `default.topo`. Perhaps it can be helpful to others too.

Closes #4025

GitOrigin-RevId: 8a03005127ba7929be4c106233217d0bdd9e0a00