xtest: add TempDir convenience method
This is more convenient than the existing `xtest.MustTempDir`, the name is chosen based on the test name.
While it is similar to the `testing.T.TempDir` method it allows to manually decide whether cleanup is desired or not.

GitOrigin-RevId: a1a98cb47a2994c5c922a86632d9a9acc7d652be
(cherry picked from commit fca5b6326d4cc30892054762ae12b8c6dd79a935)