Fix steps if down path. RawPath --> TransportPath.
When dealing with a down path SegR, the steps are always stored in the direction
of the reservation.
The steps are expected in the direction of the control plane messaging for
init confirm and init activate index.
The manager and keeper are in charge to revert the steps whenever
initializing/triggering a reservation or index.
Rename RawPath to TransportPath in SegR and its setup request.