Make reversing colibri path consistent.
Reversing means setting R to true, reverse the order of
the hop fields in the array, and swap the ingress for
the egress.
MAC computation takes R into account, and uses the correct
ingress or egress depending on it.
Adapted tests.
Hardened dataplane test to ensure that colibri reversal
is locked to these specifications.