ColibriPath and Minimal have their own Src and Dst.
Added a source and destination fields that copy the corresponding data from the SCION layer
into the colibri path. This way the colibri path is consistent and self contained, e.g. when
received from the previous AS when C=1, it will record the source and destination and pass it to
the colibri service, where both the source and destination fields will be recovered and used again.