Border routers now like our colibri packets.
Still something is wrong, as the ACK packets go from the sender to the
dispatcher and then to the receiver, but gRPC or QUIC seem not to respond.
Fix bug in slayers.SCION.SerializeTo calling SyncWithScionHeader too late.
The scionConnReader.read() method returns a net.Addr, as it must
decide between a SVCAddr or a UDPAddr.
Fix bad merge error again about setting the SRC with a DST, on
ColibriPathMinimal.SyncWithScionHeader.
Add debug messages with marker "deleteme".