fix: FABRID gRPC bugs (#173)
It is possible for the `FabridInfo` struct to be empty, e.g. when the
Daemon supports FABRID, but the control service does not support FABRID.
The checks for this were missing when looking for hops with detached
FABRID maps. This commit adds the check. Additionally, a variable name
that collides with the `fabrid` package name has been renamed.

Also fixes a bug where the modification of the snet.Path's metadata
happens on a copy. The modification of the metadata is necessary to
update the FABRID info field with fetched detached maps. In an older
version the .copy() call on the metadata was removed, this slipped back
in during a rebase.

To prevent further issues with grpc_fabrid, and considering it is one of
the only files that does not have test cases, added test cases.